refactor: update import path for Firebase auth in Login and Register components
