+++ frontend/src/pages/Register.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { createUserWithEmailAndPassword } from "firebase/auth";
-import { auth } from "../lib/firebase";
+import { auth } from "../services/firebase";
 import { useNavigate } from "react-router-dom";
 
 const Register = () => {
